fixes of review comments
